requirements.txt
fastapi
pydantic
uvicorn
watchfiles
sqlalchemy
python-dotenv